# openapi.yml
openapi: 3.0.3
info:
  title: Quick Order API
  description: |
    API documentation cho hệ thống Quick Order - Ứng dụng đặt món ăn.
    
    ## Authentication
    - Sử dụng JWT Bearer Token trong header `Authorization: Bearer <token>`
    - Refresh token được lưu trong httpOnly cookie
    
    ## Icons
    - 🔓 Public endpoint (không cần authentication)
    - 🔐 Protected endpoint (cần JWT token)
  version: 2.0.0
  contact:
    name: Quick Order Team
servers:
  - url: http://localhost:3000
    description: Development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Restaurant:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        address:
          type: string
        imageUrl:
          type: string
        owner:
          type: string
        types:
          type: array
          items:
            type: string
    RestaurantType:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        imageUrl:
          type: string

    # ============ Review ============
    CreateReviewDto:
      type: object
      required:
        - rating
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
        images:
          type: array
          items:
            type: string
    Review:
      type: object
      properties:
        _id:
          type: string
        userId:
          type: string
        restaurantId:
          type: string
        rating:
          type: integer
        comment:
          type: string
        images:
          type: array
          items:
            type: string

    # ============ Menu ============
...
    CreateOrderDto:
      type: object
      required:
        - restaurantId
        - tableId
        - items
        - lat
        - long
      properties:
        restaurantId:
          type: string
          description: MongoDB ObjectId
        tableId:
          type: string
          description: MongoDB ObjectId
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/OrderItemDto'
        lat:
          type: number
        long:
          type: number
    OrderItemDto:
      type: object
      required:
        - menuItemId
        - quantity
      properties:
        menuItemId:
          type: string
        quantity:
          type: integer
          minimum: 1
        selectedOptions:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              price:
                type: number
        note:
          type: string
    Order:
      type: object
      properties:
        _id:
          type: string
        userId:
          type: string
        restaurantId:
          type: string
        tableId:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemSnapshot'
        totalAmount:
          type: number
        status:
          $ref: '#/components/schemas/OrderStatus'
    OrderItemSnapshot:
      type: object
      properties:
        menuItemId:
          type: string
        name:
          type: string
        price:
          type: number
        quantity:
          type: integer
        selectedOptions:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              price:
                type: number
        note:
          type: string
        status:
          $ref: '#/components/schemas/OrderStatus'
    OrderStatus:
      type: string
      enum: [PENDING, COOKING, COMPLETED, CANCELED]
    UpdateOrderItemStatusDto:
      type: object
      required:
        - status
      properties:
        status:
          $ref: '#/components/schemas/OrderStatus'

    # ============ Invoice ============
...
paths:
...
  /restaurants:
    get:
      tags: [Restaurants]
      summary: Lấy danh sách nhà hàng
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Thành công
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Restaurant'
    post:
      tags: [Restaurants]
      summary: Tạo nhà hàng mới
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                address:
                  type: string
                types:
                  type: array
                  items:
                    type: string
                image:
                  type: string
                  format: binary
      responses:
        '201':
          description: Tạo thành công
  /restaurants/{id}:
    get:
      tags: [Restaurants]
      summary: Lấy chi tiết nhà hàng
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Restaurant'
  /restaurants/types:
    get:
      tags: [Restaurants]
      summary: Lấy danh sách loại nhà hàng
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Thành công
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RestaurantType'
    post:
      tags: [Restaurants]
      summary: Tạo loại nhà hàng mới
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                image:
                  type: string
                  format: binary
      responses:
        '201':
          description: Tạo thành công
  /restaurants/favorites:
    get:
      tags: [Restaurants]
      summary: Lấy danh sách nhà hàng yêu thích
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Thành công
  /restaurants/favorites/{restaurantId}:
    post:
      tags: [Restaurants]
      summary: Thêm nhà hàng vào yêu thích
      security:
        - bearerAuth: []
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Thành công
    delete:
      tags: [Restaurants]
      summary: Xóa nhà hàng khỏi yêu thích
      security:
        - bearerAuth: []
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Thành công

  # ============ REVIEWS ============
  /reviews/me:
    get:
      tags: [Reviews]
      summary: Lấy đánh giá của user hiện tại
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Thành công
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Review'
  /reviews/{restaurantId}:
    get:
      tags: [Reviews]
      summary: Lấy đánh giá của nhà hàng
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Thành công
    post:
      tags: [Reviews]
      summary: Tạo đánh giá
      security:
        - bearerAuth: []
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReviewDto'
      responses:
        '201':
          description: Tạo thành công
  /reviews/{reviewId}:
    patch:
      tags: [Reviews]
      summary: Cập nhật đánh giá
      security:
        - bearerAuth: []
      parameters:
        - name: reviewId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReviewDto'
      responses:
        '200':
          description: Cập nhật thành công
    delete:
      tags: [Reviews]
      summary: Xóa đánh giá
      security:
        - bearerAuth: []
      parameters:
        - name: reviewId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Xóa thành công

  # ============ MENUS ============
...
  /orders:
    get:
      tags: [Orders]
      summary: Lấy đơn hàng (Restaurant Owner)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Thành công
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
    post:
      tags: [Orders]
      summary: Tạo đơn hàng mới
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderDto'
      responses:
        '201':
          description: Tạo thành công
  /orders/client:
    get:
      tags: [Orders]
      summary: Lấy đơn hàng của khách
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: array
            items:
              $ref: '#/components/schemas/OrderStatus'
      responses:
        '200':
          description: Thành công
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
  /orders/kitchen/{restaurantId}:
    get:
      tags: [Orders]
      summary: Lấy đơn hàng cho bếp
      security:
        - bearerAuth: []
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Thành công
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
  /orders/{id}:
    get:
      tags: [Orders]
      summary: Lấy chi tiết đơn hàng
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
  /orders/{orderId}/items/{itemId}/status:
    patch:
      tags: [Orders]
      summary: Cập nhật trạng thái món
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrderItemStatusDto'
      responses:
        '200':
          description: Cập nhật thành công

  # ============ INVOICES ============
...
  /sse/orders/{restaurantId}:
    get:
      tags: [SSE]
      summary: SSE stream đơn hàng nhà hàng
      security:
        - bearerAuth: []
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: SSE Stream
          content:
            text/event-stream:
              schema:
                type: string
...